Corrected at91sam7s port MS/TP code where APDU size > 255 caused APDU layer to not process packets.

+++ trunk/bacnet-stack/ports/at91sam7s/dlmstp.c
@@ -143,7 +143,7 @@ static uint8_t FrameType;
    and microcontroller architectures have limits as to places to
    hold contiguous memory. */
 static uint8_t *InputBuffer;
-static uint8_t InputBufferSize;
+static uint16_t InputBufferSize;
 /* Used to store the Source Address of a received frame. */
 static uint8_t SourceAddress;
 /* "This Station," the MAC address of this node. TS is generally read from a */
